Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==2.3.3

Flask-SQLAlchemy==3.0.2
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.19.0
pylint==2.3.1
pyjwt==2.8.0
six==1.12.0
typed-ast==1.5.5
Werkzeug==0.15.6
wrapt==1.11.1
Flask-Cors==3.0.8
